moved common sentences page. also working on getting state working across entire app
