+++ lib/singletons/appdata.dart
@@ -1,0 +1,9 @@
+
+
+class AppData {
+  @protected
+  String text;
+  String get currentText => text;
+}
+
+final appData = AppData();
